requirements.txt
numpy>=1.21
pandas>=1.3
scipy>=1.7
scikit-image>=0.19
matplotlib>=3.4
seaborn>=0.11
networkx>=2.6
opencv-python>=4.5
numba>=0.54
tqdm>=4.62
lifelines>=0.27
albumentations>=1.1
pillow>=8.3
scikit-learn>=1.0
h5py>=3.6
